pyproject.toml
[project]
name = "evropuvefur-api"
version = "1.0.0"
description = "Middleware + RAG backend for Evrópuvefurinn"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "pinecone[asyncio]>=5.0.0",
    "google-genai>=1.0.0",
    "slowapi>=0.1.9",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
